+++ boards/itsybitsy_m4/examples/dotstar.rs
@@ -10,11 +10,15 @@ use panic_halt as _;
 use panic_semihosting as _;
 
 use bsp::entry;
-use hal::clock::GenericClockController;
-use hal::delay::Delay;
-use hal::pac::{CorePeripherals, Peripherals};
-use hal::prelude::*;
-use hal::timer::SpinTimer;
+
+use hal::{
+    clock::GenericClockController,
+    delay::Delay,
+    pac::{CorePeripherals, Peripherals},
+    prelude::*,
+    time::MegaHertz,
+    timer::TimerCounter,
+};
 
 use smart_leds::{hsv::RGB8, SmartLedsWrite};
 
@@ -31,9 +35,17 @@ fn main() -> ! {
     );
     let mut delay = Delay::new(core.SYST, &mut clocks);
 
-    let mut pins = bsp::Pins::new(peripherals.PORT).split();
-
-    let mut rgb = bsp::dotstar_bitbang(pins.dotstar, &mut pins.port, SpinTimer::new(12));
+    let pins = bsp::Pins::new(peripherals.PORT);
+    let gclk0 = clocks.gclk0();
+    let tc2_3 = clocks.tc2_tc3(&gclk0).unwrap();
+    let mut timer = TimerCounter::tc3_(&tc2_3, peripherals.TC3, &mut peripherals.MCLK);
+    timer.start(MegaHertz(4));
+    let mut rgb = bsp::dotstar_bitbang(
+        pins.dotstar_miso.into(),
+        pins.dotstar_mosi.into(),
+        pins.dotstar_sck.into(),
+        timer,
+    );
     let off: [RGB8; 1] = [RGB8 { r: 0, g: 0, b: 0 }];
     let red: [RGB8; 1] = [RGB8 { r: 100, g: 0, b: 0 }];
     let green: [RGB8; 1] = [RGB8 { r: 0, g: 100, b: 0 }];